Adds division into modules
